Added UDP tracker support.
See http://www.bittorrent.org/beps/bep_0015.html

Adjust types for tracker response. Use unsigned variables where
appropriate.
Handle port as uint16

+++ main.go
@@ -124,7 +124,7 @@ func listenSigInt() chan os.Signal {
 	return c
 }
 
-func startLPD(torrentSessions map[string]*torrent.TorrentSession, listenPort int) (lpd *torrent.Announcer) {
+func startLPD(torrentSessions map[string]*torrent.TorrentSession, listenPort uint16) (lpd *torrent.Announcer) {
 	lpd, err := torrent.NewAnnouncer(listenPort)
 	if err != nil {
 		log.Println("Couldn't listen for Local Peer Discoveries: ", err)
